try 2

+++ python/input.py
@@ -1,7 +1,7 @@
 #name = raw_input('What is your name? ')
 #print('Hi ' + name)
 #name = raw_input('What is your name?')
-meal_cost = raw_input('How much was your meal?')
-tip = float(meal_cost) * .20
-print('You should tip ' + float(tip))
-print('Your total cost would be ' + float(meal_cost) + float(tip))
+meal_cost = float(raw_input('How much was your meal?'))
+tip = int(meal_cost) * .20
+print('You should tip ' + int(tip)
+print('Your total cost would be ' + (meal_cost) + (tip))
